+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 9.20.0 - 2021-03-21
+
+### Added
+- Dutch translation [#1612](https://github.com/orchidsoftware/platform/pull/1612)
+
+### Fixed
+- Changed id prefix for Matrix [#1629](https://github.com/orchidsoftware/platform/pull/1629)
+- Type mismatch [#1577](https://github.com/orchidsoftware/platform/pull/1577)
+
 ## 9.19.8 - 2021-02-19
 
 ### Fixed